requirements.txt
click==7.1.2
networkx==2.5.1
pyyaml==5.2
